правки активации
