+++ js/main.js
@@ -1,11 +1,11 @@
 import {showPopup} from './popup.js';
 import {generateOffer} from './mocks.js';
-import {getActiveForm, getActiveFilter} from './form.js';
+import {setAdFormEnabled} from './form.js';
+import {setmapFilterEnabled} from'./map.js';
 
 const OFFERS_COUNT = 10;
 
 const offers = new Array(OFFERS_COUNT).fill(null).map(generateOffer);
 showPopup(offers[0]);
-
-getActiveForm();
-getActiveFilter();
+setAdFormEnabled();
+setmapFilterEnabled();
